separated spec files for card
