+++ cypress/integration/v2/blocks/card/tablet-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletAdvanced } from '.'
+
+describe( 'Card Block ( Tablet Advanced )', registerTests( [ tabletAdvanced ] ) )
