+++ cypress/integration/v2/blocks/card/tablet-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { tabletStyle } from '.'
+
+describe( 'Card Block ( Tablet Style )', registerTests( [ tabletStyle ] ) )
